# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "seedHabits开发文档",
        "title": "SeedHabits",
        "contact": {
            "name": "aogo",
            "url": "localhost:8081"
        }
    },
    "paths": {
        "/api/v1/user/info/get": {
            "get": {
                "description": "查询用户信息功能",
                "summary": "用户模块",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "description": "用户登陆功能",
                "summary": "用户模块",
                "parameters": [
                    {
                        "description": "用户名与密码",
                        "name": "LoginReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dao.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dao.LoginReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}